+++ src/Result.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Pest;
 
+use NunoMaduro\Collision\Adapters\Phpunit\Support\ResultReflection;
 use PHPUnit\TestRunner\TestResult\TestResult;
 use PHPUnit\TextUI\Configuration\Configuration;
 
@@ -44,7 +45,7 @@ final class Result
             return self::SUCCESS_EXIT;
         }
 
-        if ($configuration->failOnEmptyTestSuite() && $result->numberOfTests() === 0) {
+        if ($configuration->failOnEmptyTestSuite() && ResultReflection::numberOfTests($result) === 0) {
             return self::FAILURE_EXIT;
         }
 
@@ -53,9 +54,9 @@ final class Result
                 $returnCode = self::FAILURE_EXIT;
             }
 
-            $warnings = $result->numberOfTestsWithTestTriggeredPhpunitWarningEvents()
-                + $result->numberOfTestsWithTestTriggeredWarningEvents()
-                + $result->numberOfTestsWithTestTriggeredPhpWarningEvents();
+            $warnings = $result->numberOfTestsWithTestTriggeredPhpunitWarningEvents() +
+                +count($result->warnings())
+                + count($result->phpWarnings());
 
             if ($configuration->failOnWarning() && $warnings > 0) {
                 $returnCode = self::FAILURE_EXIT;