+++ src/TestSuite.php
@@ -11,6 +11,7 @@ use Pest\Repositories\BeforeAllRepository;
 use Pest\Repositories\BeforeEachRepository;
 use Pest\Repositories\SnapshotRepository;
 use Pest\Repositories\TestRepository;
+use Pest\Support\Str;
 use PHPUnit\Framework\TestCase;
 
 /**
@@ -104,5 +105,29 @@ final class TestSuite
         }
 
         return self::$instance;
+    }
+
+    public function getFilename(): string
+    {
+        assert($this->test instanceof TestCase);
+
+        return (fn () => self::$__filename)->call($this->test, $this->test::class); // @phpstan-ignore-line
+    }
+
+    public function getDescription(): string
+    {
+        assert($this->test instanceof TestCase);
+
+        $description = str_replace('__pest_evaluable_', '', $this->test->name());
+        $datasetAsString = str_replace('__pest_evaluable_', '', Str::evaluable($this->test->dataSetAsStringWithData()));
+
+        return str_replace(' ', '_', $description.$datasetAsString);
+    }
+
+    public function registerSnapshotChange(string $message): void
+    {
+        assert($this->test instanceof TestCase);
+
+        (fn (): string => $this->__snapshotChanges[] = $message)->call($this->test, $this->test::class); // @phpstan-ignore-line
     }
 }